% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easygraph2.R
\name{f.graph}
\alias{f.graph}
\title{create line graph}
\usage{
f.graph(data, time, tre, meth)
}
\arguments{
\item{data}{use data}

\item{time}{experiment period}

\item{tre}{treatment}

\item{meth}{measurement}

\item{ylab}{y label name}
}
\value{
graph
}
\description{
\code{f.graph} linegraph
}
